Extract keys to their own file
