+++ internal/quickstart/choose_sdk.go
@@ -18,75 +18,6 @@ var (
 	selectedSdkItemStyle = lipgloss.NewStyle().PaddingLeft(2).Foreground(lipgloss.Color("170"))
 )
 
-type listKeyMap struct {
-	Back          key.Binding
-	CloseFullHelp key.Binding
-	CursorDown    key.Binding
-	CursorUp      key.Binding
-	GoToEnd       key.Binding
-	GoToStart     key.Binding
-	NextPage      key.Binding
-	PrevPage      key.Binding
-	Quit          key.Binding
-	ShowFullHelp  key.Binding
-}
-
-func (k listKeyMap) FullHelp() [][]key.Binding {
-	return [][]key.Binding{
-		{k.CursorUp, k.CursorDown, k.PrevPage, k.NextPage},
-		{k.Back, k.Quit, k.CloseFullHelp},
-	}
-}
-
-func (k listKeyMap) ShortHelp() []key.Binding {
-	return []key.Binding{k.Back, k.Quit, k.ShowFullHelp}
-}
-
-func chooseSDKModelKeys() listKeyMap {
-	return listKeyMap{
-		Back: key.NewBinding(
-			key.WithKeys("esc"),
-			key.WithHelp("esc", "back"),
-		),
-		CursorUp: key.NewBinding(
-			key.WithKeys("up", "k"),
-			key.WithHelp("↑/k", "up"),
-		),
-		CursorDown: key.NewBinding(
-			key.WithKeys("down", "j"),
-			key.WithHelp("↓/j", "down"),
-		),
-		PrevPage: key.NewBinding(
-			key.WithKeys("left", "h", "pgup", "b", "u"),
-			key.WithHelp("←/h/pgup", "prev page"),
-		),
-		NextPage: key.NewBinding(
-			key.WithKeys("right", "l", "pgdown", "f", "d"),
-			key.WithHelp("→/l/pgdn", "next page"),
-		),
-		GoToStart: key.NewBinding(
-			key.WithKeys("home", "g"),
-			key.WithHelp("g/home", "go to start"),
-		),
-		GoToEnd: key.NewBinding(
-			key.WithKeys("end", "G"),
-			key.WithHelp("G/end", "go to end"),
-		),
-		ShowFullHelp: key.NewBinding(
-			key.WithKeys("?"),
-			key.WithHelp("?", "more"),
-		),
-		CloseFullHelp: key.NewBinding(
-			key.WithKeys("?"),
-			key.WithHelp("?", "close help"),
-		),
-		Quit: key.NewBinding(
-			key.WithKeys("ctrl+c"),
-			key.WithHelp("ctrl+c", "quit"),
-		),
-	}
-}
-
 const (
 	clientSideSDK = "client"
 	serverSideSDK = "server"
