+++ internal/quickstart/container.go
@@ -157,45 +157,3 @@ func (m ContainerModel) View() string {
 
 	return out
 }
-
-type keyMap struct {
-	Back  key.Binding
-	Enter key.Binding
-	Quit  key.Binding
-	Tab   key.Binding
-	// Help  key.Binding
-}
-
-func (k keyMap) FullHelp() [][]key.Binding {
-	return [][]key.Binding{
-		// {k.Up, k.Down, k.Left, k.Right}, // first column
-		// {k.Back, k.Quit}, // second column
-	}
-}
-
-func (k keyMap) ShortHelp() []key.Binding {
-	return []key.Binding{k.Back, k.Quit}
-}
-
-var keys = keyMap{
-	Back: key.NewBinding(
-		key.WithKeys("esc"),
-		key.WithHelp("esc", "go back"),
-	),
-	Enter: key.NewBinding(
-		key.WithKeys("enter"),
-		key.WithHelp("enter", "select"),
-	),
-	// Help: key.NewBinding(
-	// 	key.WithKeys("?"),
-	// 	key.WithHelp("?", "help"),
-	// ),
-	Quit: key.NewBinding(
-		key.WithKeys("ctrl+c"),
-		key.WithHelp("ctrl+c", "quit"),
-	),
-	Tab: key.NewBinding(
-		key.WithKeys("tab"),
-		key.WithHelp("tab", "toggle"),
-	),
-}
